requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
websockets>=11.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
pytest>=7.3.1
pytest-asyncio>=0.21.0
aiohttp>=3.8.4
asyncio>=3.4.3
python-multipart>=0.0.6
jinja2>=3.1.2
markdown>=3.4.3
